"Fix: time, distance util 개발"

+++ backend/src/main/java/com/addShot/zoosum/domain/animal/dto/response/UserAnimalDetailResponse.java
@@ -21,7 +21,7 @@ public class UserAnimalDetailResponse {
 	private int hour;
 	private int minute;
 	private int second;
-	private int lengthTogether;
+	private double lengthTogether;
 	private String fileUrl;
 
 }
